legacy/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-unsupervised"
version = "0.1.0"
description = "Unsupervised machine learning project"
authors = [
  { name="elmi Ghanmi", email="helmirealmadrid@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
  "numpy",
  "pandas",
  "scikit-learn",
  "umap-learn",
  "matplotlib",
  "seaborn",
  "pyyaml",
  "hdbscan",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
